Starting to refactor the example to work with PCC.
