+++ websocket/bot.py
